Create and finish external css for sinigang.html
Removed the home href on the bottom of the sinigang.html, as I couldn't get it to function how I want it when css was linked. A huge gap or space was seperating it from the ingredients and instructions part. I also realized that it was not necessary as it could hardly be seen; navigating the mouse cursor to click it also required more wrist movement.